requirements.txt
pandas
plotly
tensorflow
scikit-learn
xgboost
streamlit
nbformat
fastapi
uvicorn
pydantic
python-multipart
boto3
botocore
python-dotenv